If this builds/runs in CI too I'd be so happy
