+++ src/cascadia/LocalTests_TerminalApp/KeyBindingsTests.cpp
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-// #include "precomp.h"
 #include "pch.h"
 
 #include "../TerminalApp/ColorScheme.h"
