+++ src/cascadia/TerminalApp/Tab.h
@@ -2,7 +2,6 @@
 // Licensed under the MIT license.
 
 #pragma once
-// #include <winrt/Microsoft.UI.Xaml.Controls.h>
 #include "Pane.h"
 
 class Tab
